+++ packages/h3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/h3
 
+## 0.1.96
+
+### Patch Changes
+
+- @vercel/node@5.8.11
+
 ## 0.1.95
 
 ### Patch Changes