+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/nestjs
 
+## 0.2.91
+
+### Patch Changes
+
+- @vercel/node@5.8.11
+
 ## 0.2.90
 
 ### Patch Changes